CARDS-1982: Add readOnly entryMode - Rename referenceQuestion.jsx to ReadOnlyQuestion.jsx - Create new ReferenceQustion.jsx that registers a readOnlyQuestion for reference entryModes
